requirements.txt
torch
torchvision
gradio
opencv-python-headless
numpy
Pillow
